Update and refactor integration tests
