--- aws/api-gateway/cors-setup/test-service/handler.js
@@ -1,6 +1,0 @@
-'use strict';
-
-// Your first function handler
-module.exports.hello = (event, context, cb) => cb(null,
-  { message: 'Hello from API Gateway!', event }
-);
