+++ aws/api-gateway/integration-lambda/api-keys-setup/tests.js
@@ -8,7 +8,7 @@ const AWS = require('aws-sdk');
 const _ = require('lodash');
 const fetch = require('node-fetch');
 
-const Utils = require('../../../utils/index');
+const Utils = require('../../../../utils/index');
 
 const CF = new AWS.CloudFormation({ region: 'us-east-1' });
 const APIG = new AWS.APIGateway({ region: 'us-east-1' });
