+++ aws/api-gateway/integration-lambda/cors-setup/test-service/handler.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports.hello = (event, context, callback) => {
+  callback(null, { message: 'Hello from API Gateway!', event });
+};
