+++ aws/api-gateway/integration-lambda/custom-authorizers-setup/test-service/serverless.yml
@@ -3,8 +3,6 @@ service: aws-nodejs # NOTE: update this with your service name
 provider:
   name: aws
   runtime: nodejs4.3
-  apiKeys:
-    - myApiKey
 
 functions:
   hello:
@@ -13,4 +11,7 @@ functions:
       - http:
           path: hello
           method: GET
-          private: true
+          integration: lambda
+          authorizer: auth
+  auth:
+    handler: handler.auth
