+++ aws/general/custom-resources/test-service/handler.js
@@ -1,6 +1,5 @@
 'use strict';
 
-// Your first function handler
-module.exports.hello = (event, context, cb) => {
-  cb(null, { message: 'Go Serverless v1.0! Your function executed successfully!', event });
+module.exports.hello = (event, context, callback) => {
+  callback(null, { message: 'Go Serverless v1.0! Your function executed successfully!', event });
 };
