+++ aws/schedule/multipleSchedules-multipleFunctions/test-service/handler.js
@@ -1,13 +1,13 @@
 'use strict';
 
-module.exports.hello = (event, context, cb) => {
+module.exports.hello = (event, context, callback) => {
   process.stdout.write(event.source);
   process.stdout.write(event['detail-type']);
-  cb(null, { message: 'Hello from Schedule!', event });
+  callback(null, { message: 'Hello from Schedule!', event });
 };
 
-module.exports.world = (event, context, cb) => {
+module.exports.world = (event, context, callback) => {
   process.stdout.write(event.source);
   process.stdout.write(event['detail-type']);
-  cb(null, { message: 'Hello from Schedule!', event });
+  callback(null, { message: 'Hello from Schedule!', event });
 };
