+++ tests.js
@@ -6,10 +6,10 @@ require('./aws/general/deploy-invoke-remove-lifecycle/tests');
 require('./aws/general/custom-resources/tests');
 
 // API Gateway
-require('./aws/api-gateway/simple-api-setup/tests');
-require('./aws/api-gateway/custom-authorizers-setup/tests');
-require('./aws/api-gateway/cors-setup/tests');
-require('./aws/api-gateway/api-keys-setup/tests');
+require('./aws/api-gateway/integration-lambda/simple-api-setup/tests');
+require('./aws/api-gateway/integration-lambda/custom-authorizers-setup/tests');
+require('./aws/api-gateway/integration-lambda/cors-setup/tests');
+require('./aws/api-gateway/integration-lambda/api-keys-setup/tests');
 
 // Schedule
 require('./aws/schedule/multipleSchedules-multipleFunctions/tests');
